Eliminate support for Broccolifile.js
The idea was to preprocess dependencies' source files (like .coffee) on
the fly. But this is fundamentally overkill, as the correct solution
tends to be publishing precompiled files through the package manager.

Bower doesn't support pushing tarballs yet, but I don't think there's
much debate that this is a desirable feature.
https://github.com/bower/bower/issues/249