Replace bower [large] dependency with bower-config [small] dependency
Currently, the only reason Broccoli has a dependency on Bower is to properly read an interpret a `bower.json` file. The trouble is, depending on `bower` means `broccoli` ends up with __191__ npm package dependencies!

This can be improved by replacing the `bower` dependency with `bower-config`, which is used internally in `bower` here:
https://github.com/bower/bower/blob/master/lib/config.js#L3

This change mimics this usage of `bower-config` and reduces the dependencies down to 77.

Relates to #26