Fixes a bug where modules with main and lib set would get included twice
This happened because bower modules with an explicit "main" entry
always get added. If the bower module also happens to have a "lib"
folder, its files would get included twice resulting in a crash in
fs.writeFileSync.

See https://github.com/gf3/moment-range for an example of a crashing
dependency.