Added broccoli-absurd-filter to readme
I added the plugin I created for broccoli that allows it to use absurdjs( https://github.com/krasimir/absurd ) as a preprocessor for CSS or HTML files. -filter is appended as it uses broccoli-filter for one to one whereas idealistically there will be a broccoli-absurd that can do one to two. For one component.js to a component.css and component.html.