[BUGFIX] no longer cause spurious unhandled promise rejections warnings
`watcher.currentBuild` is a public API, allowing external consumers to consider the latest builds outcome. For example the broccoli express middleware will on each request follow `watcher.currentBuild`. On fulfillment it will serve the file in question, or on rejection it will provide a useful error page.

`watcher.currentBuild` is the downstream promise of several internal watcher promises, each of which already handle rejections correctly appropriately.

This commit ensures external consumers of `currentBuild` are correctly notified of the final promise state, but without any external consumers rejections of `currentBuild` are considered handled.