add the ability to ignore absolute paths from watcher
This adds an option to mark certain absolute paths as ignored by the watcher.

The motivation here is that if you want to be able to rebuild arbitrary node packages (which can put their source code at the top level of their project if they want to), and those packages emit build artifacts into some of their own subdirectories (like `dist` as used by ember-cli), you end up watching your own output.

In this situation, it is straightforward to know which path(s) are your own output, and tell broccoli not to watch them.

`sane`'s `ignored` option is perfect for this, except that it's always relative to the watchedDir, so broccoli's Watcher must translate.